refactor(api): 排队位置端点拆分至 session_queue 模块守住 1000 行上限
